feat(downloads): add option to apply parent category permissions
In the create categories view, adds the option to to take over permissions from
the parent category. This is activated as soon as a parent category is selected.
If this option is activated, the individual permissions from the new category
are hidden. If the option is deactivated, the areas are displayed again and the
new category's own permissions are taken over.